+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2024-11-15
 ### Added
 - JavaScript implementation ([#42](https://github.com/cucumber/junit-xml-formatter/pull/42))
 - Update dependency io.cucumber:messages up to v26 ([#38](https://github.com/cucumber/junit-xml-formatter/pull/38))
@@ -40,7 +41,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Java implementation ([#3](https://github.com/cucumber/junit-xml-formatter/pull/3) M.P. Korstanje)
 
-[Unreleased]: https://github.com/cucumber/junit-xml-formatter/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/cucumber/junit-xml-formatter/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/cucumber/junit-xml-formatter/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/cucumber/junit-xml-formatter/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/cucumber/junit-xml-formatter/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/cucumber/junit-xml-formatter/compare/v0.2.1...v0.3.0